Rename: 테스트용 애노테이션을 더 적절한 test 패키지 아래로 이동
