+++ src/test/java/com/patientpal/backend/test/annotation/WithCustomMockUserPatient.java
@@ -1,0 +1,15 @@
+package com.patientpal.backend.test.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.security.test.context.support.WithMockUser;
+
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Inherited
+@WithMockUser(username = "patient", roles = "USER")
+public @interface WithCustomMockUserPatient {
+}
